requirements.txt
# Core ML dependencies
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
peft>=0.6.0
accelerate>=0.24.0

# Evaluation metrics
evaluate>=0.4.0
rouge-score>=0.1.2
bert-score>=0.3.13
scikit-learn>=1.3.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0

# Deduplication and similarity
datasketch>=1.5.9
sentence-transformers>=2.2.0

# Utilities
pyyaml>=6.0
jsonlines>=4.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Optional (for better reproducibility)
safetensors>=0.4.0
gitpython>=3.1.0
